Make axes spines drawn last by default
I was previously drawing the grid after the plot was initialized, but
since the axes frame (i.e., spine) is lighter than the grid lines, the
grid lines that are at the axis limits were being drawn on top of the
spines which was ugly.
This fixes that by telling it to draw the axes spines last (read: over
the grid lines) by default.